fix: add error handling to fromUuid calls, localize loading text, and deduplicate types
- Add try/catch around fromUuid() in ExpandableDocumentList state to
  prevent unhandled promise rejections from bad UUIDs
- Replace hardcoded 'Loading...' with localize('NIMBLE.ui.loading')
- Extract SubclassChoice/EpicBoonChoice into LevelUpChoices.d.ts to
  eliminate duplication and avoid circular dependency
- Expand EPIC_BOON_LEVEL comment to clarify Nimble rule
