+++ src/view/dialogs/const/levelUpConstants.ts
@@ -1,5 +1,8 @@
 /** Subclass is chosen when leveling to this class level */
 export const SUBCLASS_LEVEL = 3;
 
-/** Epic boon is chosen when leveling to this class level (Nimble class rule) */
+/**
+ * Epic boon is chosen when leveling to this class level.
+ * Per Nimble's rules, all classes gain an epic boon at level 19.
+ */
 export const EPIC_BOON_LEVEL = 19;
