+++ types/components/CharacterLevelUpDialog.d.ts
@@ -3,23 +3,11 @@ import type { NimbleFeatureItem } from '#documents/item/feature.js';
 import type { ExpandableDocumentItem } from './ExpandableDocumentList.d.ts';
 import type GenericDialog from '#documents/dialogs/GenericDialog.svelte.js';
 
+export type { SubclassChoice, EpicBoonChoice } from './LevelUpChoices.d.ts';
+
 export interface CharacterLevelUpDialogProps {
 	document: NimbleCharacter;
 	dialog: GenericDialog;
-}
-
-export interface SubclassChoice {
-	uuid: string;
-	name: string;
-	img: string;
-	system: { parentClass: string };
-}
-
-export interface EpicBoonChoice {
-	uuid: string;
-	name: string;
-	img: string;
-	system: { boonType: string; description: string };
 }
 
 export interface LevelUpSubmitData {
