+++ types/components/LevelUpChoices.d.ts
@@ -1,0 +1,13 @@
+export interface SubclassChoice {
+	uuid: string;
+	name: string;
+	img: string;
+	system: { parentClass: string };
+}
+
+export interface EpicBoonChoice {
+	uuid: string;
+	name: string;
+	img: string;
+	system: { boonType: string; description: string };
+}
